Moved the rating and review section to the secreens_review file and replaced the Textstyle parts with ThemeStyle here
