+++ lib/src/features/products/presentation/screens/screen_product_details.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_boilerplate_code/src/routes/routes.dart';
 
 class ScreenProductDetails extends StatelessWidget {
   const ScreenProductDetails({super.key});
@@ -246,19 +247,29 @@ class ScreenProductDetails extends StatelessWidget {
                               style: TextStyle(
                                   color: Colors.grey[600], fontSize: 14)),
                           SizedBox(height: 4),
-                          Row(
-                            children: const [
-                              Icon(Icons.star, color: Colors.orange, size: 16),
-                              Icon(Icons.star, color: Colors.orange, size: 16),
-                              Icon(Icons.star, color: Colors.orange, size: 16),
-                              Icon(Icons.star, color: Colors.orange, size: 16),
-                              Icon(Icons.star, color: Colors.orange, size: 16),
-                              SizedBox(width: 6),
-                              Text("(7)",
-                                  style: TextStyle(
-                                      color: Colors.black87, fontSize: 14)),
-                            ],
-                          ),
+                          InkWell(
+                            onTap: () {
+                              Navigator.pushNamed(context, Routes.reviews);
+                            },
+                            child: Row(
+                              children: const [
+                                Icon(Icons.star,
+                                    color: Colors.orange, size: 16),
+                                Icon(Icons.star,
+                                    color: Colors.orange, size: 16),
+                                Icon(Icons.star,
+                                    color: Colors.orange, size: 16),
+                                Icon(Icons.star,
+                                    color: Colors.orange, size: 16),
+                                Icon(Icons.star,
+                                    color: Colors.orange, size: 16),
+                                SizedBox(width: 6),
+                                Text("(7)",
+                                    style: TextStyle(
+                                        color: Colors.black87, fontSize: 14)),
+                              ],
+                            ),
+                          )
                         ],
                       ),
 
@@ -298,234 +309,6 @@ class ScreenProductDetails extends StatelessWidget {
             ),
 
             const SizedBox(height: 20),
-
-            // rating section
-            Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Padding(
-                  padding: EdgeInsets.symmetric(vertical: 16, horizontal: 20),
-                  child: Text(
-                    "Ratings & Reviews",
-                    style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                  ),
-                ),
-                Row(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    // part 1 average rating
-                    Column(
-                      children: const [
-                        Text(
-                          "4.3",
-                          style: TextStyle(
-                            fontSize: 32,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        SizedBox(height: 4),
-                        Text("23 ratings"),
-                      ],
-                    ),
-                    const SizedBox(width: 40),
-
-                    // part 2 star labels
-                    Column(
-                      crossAxisAlignment: CrossAxisAlignment.end,
-                      children: List.generate(5, (index) {
-                        int filledStars = 5 - index;
-                        return Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          mainAxisSize: MainAxisSize.min,
-                          children: List.generate(
-                            filledStars,
-                            (_) => const Icon(Icons.star,
-                                size: 20, color: Colors.orange),
-                          ),
-                        );
-                      }),
-                    ),
-
-                    const SizedBox(width: 10),
-                    // part 3 bar graph
-                    Column(
-                      children: [
-                        _buildRatingBar(0.9),
-                        _buildRatingBar(0.7),
-                        _buildRatingBar(0.4),
-                        _buildRatingBar(0.2),
-                        _buildRatingBar(0.1),
-                      ],
-                    ),
-
-                    const SizedBox(width: 40),
-
-                    // Part 4: Rating count
-                    Column(
-                      mainAxisAlignment: MainAxisAlignment.start,
-                      children: const [
-                        Text("8"),
-                        Text("11"),
-                        Text("0"),
-                        Text("0"),
-                        Text("2"),
-                      ],
-                    )
-                  ],
-                ),
-              ],
-            ),
-
-            // review title section
-            Padding(
-              padding: const EdgeInsets.symmetric(vertical: 16.0),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text(
-                    "8 reviews",
-                    style: TextStyle(fontSize: 14, fontWeight: FontWeight.w500),
-                  ),
-                  Row(
-                    children: [
-                      Checkbox(
-                        value: false,
-                        onChanged: (bool? newValue) {},
-                      ),
-                      const Text(
-                        "With Photo",
-                        style: TextStyle(fontSize: 14),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-
-            // the review section
-            Container(
-              margin: const EdgeInsets.symmetric(vertical: 12, horizontal: 12),
-              child: Stack(
-                children: [
-                  // review card
-                  Container(
-                    padding: const EdgeInsets.all(16),
-                    margin: const EdgeInsets.only(left: 24, top: 12),
-                    decoration: BoxDecoration(
-                      color: Colors.grey[100],
-                      borderRadius: BorderRadius.circular(12),
-                      border: Border.all(color: Colors.grey.shade300),
-                    ),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const SizedBox(height: 8),
-
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            // name and rating
-                            Expanded(
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  const Text(
-                                    "Jane Doe",
-                                    style: TextStyle(
-                                      fontWeight: FontWeight.bold,
-                                      fontSize: 14,
-                                    ),
-                                  ),
-                                  const SizedBox(height: 4),
-                                  Row(
-                                    children: List.generate(5, (index) {
-                                      return const Icon(
-                                        Icons.star,
-                                        size: 16,
-                                        color: Colors.orange,
-                                      );
-                                    }),
-                                  ),
-                                ],
-                              ),
-                            ),
-                            // Date
-                            const Text(
-                              "May 13, 2025",
-                              style: TextStyle(
-                                color: Colors.grey,
-                                fontSize: 12,
-                              ),
-                            ),
-                          ],
-                        ),
-
-                        const SizedBox(height: 12),
-
-                        // Review Description
-                        const Text(
-                          "The product is really good and fits perfectly. The material is soft and comfortable to wear. Highly recommended!",
-                          style: TextStyle(fontSize: 14),
-                        ),
-
-                        const SizedBox(height: 12),
-
-                        // helpful icon part
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.end,
-                          children: [
-                            const Text(
-                              "Helpful",
-                              style: TextStyle(
-                                fontSize: 14,
-                                color: Colors.grey,
-                              ),
-                            ),
-                            const SizedBox(width: 4),
-                            Icon(
-                              Icons.thumb_up_alt_outlined,
-                              size: 18,
-                              color: Colors.grey[600],
-                            ),
-                          ],
-                        ),
-                      ],
-                    ),
-                  ),
-
-                  // circular profile image (overlapping top-left corner)
-                  Positioned(
-                    top: 0,
-                    left: 0,
-                    child: CircleAvatar(
-                      radius: 20,
-                      backgroundColor: Colors.white,
-                      backgroundImage: AssetImage('assets/images/image 2.png'),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-
-            // write a review section
-            Align(
-              alignment: Alignment.centerRight,
-              child: TextButton.icon(
-                onPressed: () {},
-                icon: const Icon(Icons.edit, size: 18),
-                label: const Text(
-                  "Write a Review",
-                  style: TextStyle(fontSize: 14),
-                ),
-                style: TextButton.styleFrom(
-                  foregroundColor: Colors.white,
-                  backgroundColor: Colors.red,
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 20, vertical: 6),
-                  textStyle: const TextStyle(fontWeight: FontWeight.bold),
-                ),
-              ),
-            ),
           ],
         ),
       ),
@@ -550,22 +333,4 @@ class ScreenProductDetails extends StatelessWidget {
       ),
     );
   }
-}
-
-Widget _buildRatingBar(double value) {
-  return Container(
-    margin: const EdgeInsets.symmetric(vertical: 2),
-    width: 100,
-    height: 15,
-    child: FractionallySizedBox(
-      alignment: Alignment.centerLeft,
-      widthFactor: value,
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.orange,
-          borderRadius: BorderRadius.circular(5),
-        ),
-      ),
-    ),
-  );
 }
