+++ lib/src/routes/routes.dart
@@ -1,4 +1,5 @@
 class Routes {
   static const String homeScreen = "/";
   static const String productDetails = "/productDetails";
+  static const String reviews = "/reviews";
 }
